Move expr_cast to a separate file
